Implement approver 'role' in setReadyForReviewOnPublishAction
